merge

+++ xutils/src/main/java/org/xutils/common/util/ProcessLock.java
@@ -14,7 +14,9 @@ import java.io.IOException;
 import java.nio.channels.FileChannel;
 import java.nio.channels.FileLock;
 import java.text.DecimalFormat;
-import java.util.HashMap;
+import java.util.Iterator;
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 
 /**
  * 进程间锁, 仅在同一个应用中有效.
@@ -25,20 +27,26 @@ public final class ProcessLock implements Closeable {
     private final FileLock mFileLock;
     private final File mFile;
     private final Closeable mStream;
+    private final boolean mWriteMode;
 
     private final static String LOCK_FILE_DIR = "process_lock";
-    private final static int PID = android.os.Process.myPid();
-    private final static HashMap<String, ProcessLock> LOCK_MAP = new HashMap<String, ProcessLock>(5);
+    /**
+     * key1: lockName
+     * key2: fileLock.hashCode()
+     */
+    private final static DoubleKeyValueMap<String, Integer, ProcessLock> LOCK_MAP = new DoubleKeyValueMap<String, Integer, ProcessLock>();
 
     static {
-        IOUtil.deleteFileOrDir(x.app().getDir(LOCK_FILE_DIR, Context.MODE_PRIVATE));
+        File dir = x.app().getDir(LOCK_FILE_DIR, Context.MODE_PRIVATE);
+        IOUtil.deleteFileOrDir(dir);
     }
 
-    private ProcessLock(String lockName, File file, FileLock fileLock, Closeable stream) {
+    private ProcessLock(String lockName, File file, FileLock fileLock, Closeable stream, boolean writeMode) {
         mLockName = lockName;
         mFileLock = fileLock;
         mFile = file;
         mStream = stream;
+        mWriteMode = writeMode;
     }
 
     /**
@@ -60,7 +68,7 @@ public final class ProcessLock implements Closeable {
      * @param maxWaitTimeMillis 最大值 1000 * 60
      * @return null 或 进程锁, 如果锁已经被占用, 则在超时时间内继续尝试获取该锁.
      */
-    public static ProcessLock tryLock(final String lockName, final boolean writeMode, final long maxWaitTimeMillis) {
+    public static ProcessLock tryLock(final String lockName, final boolean writeMode, final long maxWaitTimeMillis) throws InterruptedException {
         ProcessLock lock = null;
         long expiryTime = System.currentTimeMillis() + maxWaitTimeMillis;
         String hash = customHash(lockName);
@@ -71,6 +79,8 @@ public final class ProcessLock implements Closeable {
             } else {
                 try {
                     Thread.sleep(1); // milliseconds
+                } catch (InterruptedException iex) {
+                    throw iex;
                 } catch (Throwable ignored) {
                 }
             }
@@ -111,17 +121,23 @@ public final class ProcessLock implements Closeable {
         synchronized (LOCK_MAP) {
             if (fileLock != null) {
                 try {
-                    LOCK_MAP.remove(lockName);
-                    fileLock.release();
-                    LogUtil.d("released: " + lockName + ":" + PID);
-                } catch (Throwable ignored) {
+                    LOCK_MAP.remove(lockName, fileLock.hashCode());
+                    ConcurrentHashMap<Integer, ProcessLock> locks = LOCK_MAP.get(lockName);
+                    if (locks == null || locks.isEmpty()) {
+                        IOUtil.deleteFileOrDir(file);
+                    }
+
+                    if (fileLock.channel().isOpen()) {
+                        fileLock.release();
+                    }
+                } catch (Throwable ex) {
+                    LogUtil.e(ex.getMessage(), ex);
                 } finally {
                     IOUtil.closeQuietly(fileLock.channel());
                 }
             }
 
             IOUtil.closeQuietly(stream);
-            IOUtil.deleteFileOrDir(file);
         }
     }
 
@@ -141,46 +157,51 @@ public final class ProcessLock implements Closeable {
     private static ProcessLock tryLockInternal(final String lockName, final String hash, final boolean writeMode) {
         synchronized (LOCK_MAP) {
 
-            // android对文件锁共享支持的不好, 暂时全部互斥.
-            if (LOCK_MAP.containsKey(lockName)) {
-                ProcessLock lock = LOCK_MAP.get(lockName);
-                if (lock == null) {
-                    LOCK_MAP.remove(lockName);
-                } else if (lock.isValid()) {
-                    return null;
-                } else {
-                    LOCK_MAP.remove(lockName);
-                    lock.release();
+            ConcurrentHashMap<Integer, ProcessLock> locks = LOCK_MAP.get(lockName);
+            if (locks != null && !locks.isEmpty()) {
+                Iterator<Map.Entry<Integer, ProcessLock>> itr = locks.entrySet().iterator();
+                while (itr.hasNext()) {
+                    Map.Entry<Integer, ProcessLock> entry = itr.next();
+                    ProcessLock value = entry.getValue();
+                    if (value != null) {
+                        if (!value.isValid()) {
+                            itr.remove();
+                        } else if (writeMode) {
+                            return null;
+                        } else if (value.mWriteMode) {
+                            return null;
+                        }
+                    } else {
+                        itr.remove();
+                    }
                 }
             }
 
-            FileInputStream in = null;
-            FileOutputStream out = null;
-            Closeable stream = null;
             FileChannel channel = null;
+            Closeable stream = null;
             try {
                 File file = new File(
                         x.app().getDir(LOCK_FILE_DIR, Context.MODE_PRIVATE),
                         hash);
                 if (file.exists() || file.createNewFile()) {
+
                     if (writeMode) {
-                        out = new FileOutputStream(file, false);
+                        FileOutputStream out = new FileOutputStream(file, false);
                         channel = out.getChannel();
                         stream = out;
                     } else {
-                        in = new FileInputStream(file);
+                        FileInputStream in = new FileInputStream(file);
                         channel = in.getChannel();
                         stream = in;
                     }
                     if (channel != null) {
                         FileLock fileLock = channel.tryLock(0L, Long.MAX_VALUE, !writeMode);
                         if (isValid(fileLock)) {
-                            LogUtil.d("lock: " + lockName + ":" + PID);
-                            ProcessLock processLock = new ProcessLock(lockName, file, fileLock, stream);
-                            LOCK_MAP.put(lockName, processLock);
-                            return processLock;
+                            ProcessLock result = new ProcessLock(lockName, file, fileLock, stream, writeMode);
+                            LOCK_MAP.put(lockName, fileLock.hashCode(), result);
+                            return result;
                         } else {
-                            release(lockName, fileLock, file, out);
+                            release(lockName, fileLock, file, stream);
                         }
                     } else {
                         throw new IOException("can not get file channel:" + file.getAbsolutePath());
@@ -188,8 +209,7 @@ public final class ProcessLock implements Closeable {
                 }
             } catch (Throwable ignored) {
                 LogUtil.d("tryLock: " + lockName + ", " + ignored.getMessage());
-                IOUtil.closeQuietly(in);
-                IOUtil.closeQuietly(out);
+                IOUtil.closeQuietly(stream);
                 IOUtil.closeQuietly(channel);
             }
         }
@@ -199,7 +219,7 @@ public final class ProcessLock implements Closeable {
 
     @Override
     public String toString() {
-        return mLockName;
+        return mLockName + ": " + mFile.getName();
     }
 
     @Override
